Hoist interaction utils further for better re-use
Since I first did this, I've ended up using more of the aspects
of the dpeloy logic elsewhere.
